Implement nasty fucking hack to be able to properly test the server
This is the worst code I have ever written for a test.

I would like to thank asyncio and Python for having such a shitty
fucking thread model that this is necessary. I would also like to thank
aiohttp for not only being the only python-based web server not based on
gunicorn and that shitty standard[^1], but using asyncio to the point
that making tests outside their very specific testing scenario is
absolutely impossible without resorting to not one, but two inexplicable
and very different stalling tactics needed to beat asyncio into
submission.

I would further like to thank the pytest-aiohttp devs for not explicitly
stating that the server would, in fact, not run without the main thread
yielding so the asyncio loop can actually process the requests, and
that the library actually needs the TestClient to work because asyncio
is a piece of shit that doesn't multithread.
This little manoeuvre cost me 51 years.

The documentation for both pytest-asyncio and pytest-aiohttp is among
the worst documentation I've had the misfortune of reading - and I've
implemented shit using libarchive, and libarchive's documentation is
already so bad I did not expect it could get worse in widely used
production-ready code.

Is this partly my fault for not understanding asyncio in the first
place? Absolutely. I'm still going to blame aiohttp for making the
server such a pain in the ass to test outside the very specific bounds
of their specially written test code, and asyncio for being a pain in
the ass in general.

I initially believed I couldn't use threads at all, but that's because
my initial attempt at using threads used Thread + `web._run_app`, which
apparently _requires_ specific main thread features:

> ...
> E           RuntimeError: set_wakeup_fd only works in main thread

For a Thread to work, a specially written TestServer still has to be
used. This is the same class pytest-aiohttp uses under the hood, but the
class itself comes from aiohttp proper.

Having also worked with ktor recently (Kotlin), which uses a very
similar setup for tests, _fucking why_? As a utility, I understand it,
but when I work with crowcpp, it spawns its own thread, and then fucks
off. There's two function calls to start the server, one of which is
blocking, one of which isn't. You want to return control of the main
thread in tests so you can test shit? No fucking problem, have a 2 line
function (excl. brackets and function declaration) that wraps the sync
`run` function:
    https://github.com/CrowCpp/Crow/blob/d3038b6ac85b40795f142864370d08afcb39034e/include/crow/app.h#L588-L590
In one of my projects, I don't use the `std::future` for reasons I don't
recall, and it's instead a two-line call (excluding brackets) to create
a thread:
    runner = std::thread([this]() {
        this->serv->run();
    });

100% control over the control flow, where if shit doesn't execute, it's
because you failed to start it, or because you failed to keep it alive.
Asyncio hiding all that in an obscure loop makes it such a pain in the
ass to diagnose whether it's PEBKAC or hard limits of asyncio.

I cannot believe I'm writing this, but the threading model used by
JavaScript (a language that has no business being outside a browser)
is orders of magnitude better equipped for async code than asyncio's
setup is. Async shoehorned into a language that isn't already built
around an event loop makes for crappy async quality, and a shitty
developer experience trying to support it. Async in JS is just normal
JS with some extra keywords, while async in python is "rip out
everything, you used async in one deeply nested function and now you
have to make everything async or create a deeply nested one-off asyncio
loop that you immediately throw away".

Every time I end up trying to use asyncio, I immediately regret it
because of this. You just want to run a single server heavily based
around async without blocking your test? it needs to go in a single
test, and the overhead is a total of 24 lines of code to
* Set up the loop
* Start the thread
    * Wait for a second or nothing works for no obvious reason
    * Start the loop
* Start the server
* Wait for the server to start so the tests don't explode
* Run the tests
* Stop the server
* Stop the loop
* Force the server to stop with a web request
* Join the thread

Async has benefits and all that shit, but at this cost, I'd rather throw
more threads at the problem than suffer through this much shit just so I
can run some fucking tests.

[^1]: the only python-based webserver based on gunicorn of any
    meaningful size.